[REFACTOR]: Modularize models and schema into each app (complaints, users, gasoline)
