+++ complaints/admin.py
@@ -1,3 +1,29 @@
+""" Complaint admin"""
+
+# Django
 from django.contrib import admin
 
-# Register your models here.
+# Models
+from complaints.models import Complaint
+
+@admin.register(Complaint)
+class  ComplaintAdmin(admin.ModelAdmin):
+    """ Complaint model Admin"""
+
+    list_display = (
+        'user',
+        'station',
+        'description',
+        'link_evidence',
+        'type_complaint',
+        'date',
+        'actual_price',
+        'offered_price',
+    )
+
+    list_filter = (
+        'user',
+        'station',
+        'type_complaint',
+        'date',
+    )
