+++ complaints/models/__init__.py
@@ -1,0 +1,1 @@
+from .complaint import *
