--- gasoline/models.py
@@ -1,153 +1,0 @@
-"""Gasoline models """
-
-# Django
-from django.db import models
-from django.contrib.auth.models import User
-
-class Station(models.Model):
-    """ Station Model"""
-    # Information
-    name = models.CharField(
-        max_length=200,
-    )
-    about = models.CharField(
-        max_length=200,
-        blank=True,
-        null=True,
-        help_text='Short description of the station'
-        ) 
-    picture = models.ImageField(
-        upload_to=None,
-        blank=True,
-        null=True
-        )
-    register = models.CharField(
-        max_length=64,
-        unique=True,
-        blank=True,
-        null=True,
-        help_text='Unique Register given for the Mexican Govenment')
-    
-    # Location
-    latitude = models.FloatField(max_length=9)
-    longitude = models.FloatField(max_length=9)
-    town = models.CharField(
-        max_length=50,
-        blank=True,
-        null=True,
-        )
-    state = models.CharField(
-        max_length=50,
-        blank=True,
-        null=True,
-        help_text='Official name of the Mexican state where is located the station'
-        )
-
-    # Status
-    is_active = models.BooleanField(default=False)
-    status = models.CharField(
-        max_length=10,
-        default='ghost',
-        help_text="Station' status based on their activity. It changes when a user verify the station "
-        )
-
-    def __str__(self):
-        """Return station name"""
-        return self.name
-
-
-class Price(models.Model):
-    """ Price Model
-
-    Model that holds the prices of one type of gasoline by station.
-    """
-    station = models.ForeignKey(Station, on_delete=models.PROTECT)
-
-    GAS_CHOICES = [
-        ('premium','premium'),
-        ('regular','regular'),
-        ('diesel','diesel')
-    ]
-
-    gas_type = models.CharField(
-        max_length=20,
-        choices=GAS_CHOICES,
-        help_text='Type of gasoline between the GAS_CHOICES',
-        )
-    price = models.FloatField(
-        max_length=5,
-        )
-    date = models.DateTimeField(
-        'created at',
-        auto_now_add=True,
-        help_text='Date Time on wich the prices are registred'
-        )
-    def __str__(self):
-        """ Returns price."""
-        return f'{self.gas_type} cuesta {self.price} en {self.station}'
-
-
-class Profile(models.Model):
-    """Profile Model 
-    
-    Model that extends the user model and add extra information of the user.
-    """
-
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    
-    phone_number = models.CharField(
-        max_length=20, 
-        blank=True
-    )
-
-    def __str__(self):
-        """Returns string representation of user"""
-        return self
-
-
-class Complaint(models.Model):
-    """Complaint Model
-
-    A complaint is a bad score that an user makes 
-    about the prices in a special station.
-    """
-    user = models.ForeignKey(
-        User, 
-        on_delete=models.PROTECT
-    )
-    station = models.ForeignKey(
-        Station,
-        on_delete=models.CASCADE
-        )
-    description = models.TextField(
-        max_length=500,
-        null=False,
-        )
-    link_evidence = models.CharField(max_length=255)
-    type_complaint = models.CharField(max_length=255)
-
-    date = models.DateTimeField(
-        auto_now_add=True,
-        help_text='Datetime when the complaint is created'
-        )
-
-    # Prices
-    actual_price = models.ForeignKey(
-        Price, 
-        on_delete=models.PROTECT,
-        help_text='Price registered in the system when creating the complaint'
-        )
-    offered_price = models.FloatField(
-        max_length=5,
-        help_text='Price published in the station'
-        )
-
-    def __str__(self):
-        """ Returns user and datetime of the complaint."""
-        return '{}: Actual price {}, offered_price {}'.format(
-                self.user,
-                self.actual_price,
-                self.offered_price,
-                )
-
-
