+++ gasoline/schema.py
@@ -1,10 +1,9 @@
 """Schemas for gasoline app """
 # Django
 import django_filters
-from django.contrib.auth.models import User
 
 # Models
-from .models import Price, Station, Profile, Complaint
+from gasoline.models import Price, Station
 
 # Graphene
 import graphene
@@ -47,49 +46,7 @@ class PriceType(DjangoObjectType):
         )
 
 
-class UserType(DjangoObjectType):
-    """Type for user model"""
-    class Meta:
-        """Class Meta"""
-        model = User
-        fields = (
-            'username',
-            'first_name',
-            'last_name',
-            'email',
-        )
-
-
-class ProfileType(DjangoObjectType):
-    """ Type for Profile model"""
-    class Meta:
-        """Class Meta"""
-        model = Profile
-        fields = (
-            "user",
-            "phone_number"
-        )
-
-
-class ComplaintType(DjangoObjectType):
-    """ Type for Complaint Model"""
-    class Meta:
-        """Class Meta"""
-        model = Complaint
-        fields = (
-            'user',
-            'station',
-            'description',
-            'link_evidence',
-            'type_complaint',
-            'date',
-            'actual_price',
-            'offered_price'
-        )
-
-
 #-----------NODE-QUERIES----------
-
 class StationNode(DjangoObjectType):
     class Meta:
         model = Station
@@ -106,6 +63,7 @@ class StationNode(DjangoObjectType):
         }
         interfaces = (relay.Node,)
 
+
 class PriceNode(DjangoObjectType):
     class Meta:
         model = Price
@@ -119,60 +77,11 @@ class PriceNode(DjangoObjectType):
         interfaces = (relay.Node,)
 
 
-class ProfileNode(DjangoObjectType):
-    class Meta:
-        model = Profile
-        filter_fields = {
-            "user": ['exact'],
-            "phone_number": ['exact', 'icontains','istartswith'],
-        }
-        interfaces = (relay.Node,)
-
-
-class ComplaintNode(DjangoObjectType):
-    class Meta:
-        model = Complaint
-        filter_fields = {
-            'user': ['exact'],
-            'station__id': ['exact', 'icontains','istartswith'],
-            'description': ['exact', 'icontains','istartswith'],
-            'type_complaint': ['exact', 'icontains','istartswith'],
-            'date': ['exact', 'icontains','istartswith'],
-            'actual_price__price': ['exact', 'icontains','istartswith'],
-            'offered_price': ['exact', 'icontains','istartswith'],
-        }
-        interfaces = (relay.Node,)
-
-#-------------MUTATIONS------------
-
-class UpdateUser(graphene.Mutation):
-    class Arguments:
-        user = graphene.String(required=True)
-        password = graphene.String(required=True)
-        email = graphene.String(required=True)
-        first_name = graphene.String(required=True)
-        last_name = graphene.String()
-
-    user = graphene.Field(UserType)
-
-    def mutate(self, info, user, password,email,first_name,last_name):
-        user = User(
-                password=password,
-                email=email,
-                first_name=first_name,
-                last_name=last_name
-            )
-        user.save()
-        return UpdateUser(user=user)
-
 #---------------SCHEMA---------------
-
 class Query(graphene.ObjectType):
-    """General Query class"""
+    """Gasoline Query class"""
     all_stations = graphene.List(StationType)
     all_prices = graphene.List(PriceType)
-    all_profiles = graphene.List(ProfileType)
-    all_complaints = graphene.List(ComplaintType)
 
     def resolve_all_stations(root, info):
         """ Return all the stations """
@@ -182,29 +91,12 @@ class Query(graphene.ObjectType):
         """ Return all prices """
         return Price.objects.all()
 
-    def resolve_all_profiles(root, info):
-        """ Return all profiles"""
-        return Profile.objects.all()
-    
-    def resolve_all_complaints(root, info):
-        """ Return all complaints"""
-        return Complaint.objects.all()
-
-    """Node Query class"""
+    # Node Query class
     station = relay.Node.Field(StationNode)
     node_station = DjangoFilterConnectionField(StationNode)
 
     price = relay.Node.Field(PriceNode)
     node_price = DjangoFilterConnectionField(PriceNode)
 
-    profile = relay.Node.Field(ProfileNode)
-    node_profile = DjangoFilterConnectionField(ProfileNode)
-
-    complaint = relay.Node.Field(ComplaintNode)
-    node_complaint = DjangoFilterConnectionField(ComplaintNode)
-
-
 class Mutation(graphene.ObjectType):
-    update_user = UpdateUser.Field()
-
-schema = graphene.Schema(query=Query, mutation=Mutation)
+    """ Gasoline Mutation class."""
